Enable usage of the upstream apache module
In system-config we currently cohabitate management of the eavesdrop
site with management of the ptgbot site. We need to convert the
eavesdrop site to use the upstream apache module's custom vhost resource
for reasons given in that change[1]. The problem is we can't include
both the httpd class and the apache class because they both declare the
'httpd' package resource and cause a duplicate definition error. This
patch allows the ptgbot module to work with both classes while we
convert system-config and eavesdrop to use the upstream one.

[1] https://review.openstack.org/564904

Change-Id: I79150d1bf4ccb23070f2166580b5352db724b2dc