No longer build index page in puppet-ptgbot
Index page used to be shipped and configured in the
puppet-ptgbot module itself. Now it is provided by the
ptgbot itself and is dynamically generated from JSON
data the same way the PTG schedule page is.

This avoids the need to update puppet-ptgbot at every
event to update the links on that page. They can now be
dynamically provided using the ~fetchdb command.

Depends-On: https://review.openstack.org/626907
Change-Id: Id5c84cf3d2fcff0e772764e40dea64c1383a85bc